numpy and gluonts api update

+++ test/distributions/test_piecewise_linear.py
@@ -43,7 +43,7 @@ def empirical_cdf(
     edges = []
 
     batch_shape = samples.shape[1:]
-    agg_batch_dim = np.prod(batch_shape, dtype=np.int)
+    agg_batch_dim = np.prod(batch_shape, dtype=np.int32)
 
     samples = samples.reshape((samples.shape[0], -1))
